requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
#Agregamos el SDK de snowflake para conectar y usar con python
astro-sdk-python[snowflake]
apache-airflow-providers-snowflake
apache-airflow-providers-slack